+++ tests/agent_server/conftest.py
@@ -178,6 +178,14 @@ def resource_container():
 
         fields = ("key", "value", "purged")
 
+    @resource("test::BadEventsStatus", agent="agent", id_attribute="key")
+    class BadEventS(Resource):
+        """
+        Set `ctx.set_status(const.ResourceState.failed)` in process_events().
+        """
+
+        fields = ("key", "value", "purged")
+
     @resource("test::BadPost", agent="agent", id_attribute="key")
     class BadPostR(Resource):
         """
@@ -412,6 +420,21 @@ def resource_container():
 
         def process_events(self, ctx, resource, events):
             raise Exception()
+
+    @provider("test::BadEventsStatus", name="test_bad_events_status")
+    class BadEventStatus(ResourceHandler):
+        def check_resource(self, ctx, resource):
+            current = resource.clone()
+            return current
+
+        def do_changes(self, ctx, resource, changes):
+            pass
+
+        def can_process_events(self) -> bool:
+            return True
+
+        def process_events(self, ctx, resource, events):
+            ctx.set_status(const.ResourceState.failed)
 
     @provider("test::BadPost", name="test_bad_posts")
     class BadPost(Provider):